0.4.3 Input Dialog Service
1. Generates provider InputDialogService
2. Removes import for httpClient from input-dialog-service.ts as not needed, remove injection.
3. Adds Import InputDialogService to dashboard.ts and injection
4. Adds Import ProposalsService to input-dialog-service.ts and injection
5. Adds Import AlertController to input-dialog-service.ts and injection
6. Extracts showAddProposalPrompt() and showEditProposalPrompt() from dashboard.ts to input-dialog-service.ts
7. Adds updated call in editProposal() and addProposal() in dashboard.ts to inputDialogService
8. Renames showEditProposalPrompt() to showPrompt() in input-dialog-service.ts
9. Updates calls in dashboard.ts from editProposal() and addProposal() to point to showPrompt()
10. Adds ? to each param of showPrompt to make optional for add case.
11. Adds use of ternary null operator for showPrompt function title, message, and inputs values
12. Adds if else loop to showPrompt() for final action to if index !== undefined, then editProposal,
else addProposal in input-dialog-service.ts
13. Removes ShowAddProposalPrompt() function that was combined into showPrompt()
