+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "swdv-665-final-project",
-  "version": "0.4.2",
+  "version": "0.4.3",
   "author": "Michael G. Craft",
   "homepage": "https://github.com/SWDV-665/final-project-mcraft3",
   "private": true,
